+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ## Changed
 
+- [breaking] Support only metabase version >= `0.40.0`
+- Support the pagination feature of `/api/user` [Ref](https://github.com/metabase/metabase/wiki/What%27s-new-in-0.40.0-for-Metabase-REST-API-clients)
+
 # 0.0.24 (2022-11-13 / 23678d6)
 
 ## Added