+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 0.0.45 (2022-11-24 / f6273b8)
 
 ## Added