+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 0.0.50 (2023-01-19 / 8e058ff)
 
 ## Added