+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.0.56 (2023-04-14 / fd0bc4a)
 
 ## Added